pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narrowgate"
version = "3.0.0"
description = "窄门 NarrowGate — AI-driven soul evolution platform"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.12"
authors = [
    {name = "Dr. MoKang", email = "mokang@mokangmedical.com"},
]
keywords = ["soul-evolution", "self-improvement", "ai-coaching", "narrowgate"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Modules",
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "httpx>=0.28",
    "aiofiles>=24.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "wechatpayv3>=1.2.0",
    "alipay-sdk-python>=3.0.0",
    "PyJWT>=2.8",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.28",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/MoKangMedical/narrowgate"
Repository = "https://github.com/MoKangMedical/narrowgate"
Issues = "https://github.com/MoKangMedical/narrowgate/issues"

# ============================================================
# Pytest
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]

# ============================================================
# Coverage
# ============================================================
[tool.coverage.run]
source = ["src"]
omit = [
    "src/ui/*",
    "tests/*",
    "scripts/*",
]

[tool.coverage.report]
show_missing = true
fail_under = 40
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ============================================================
# Ruff (linter + formatter)
# ============================================================
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["core", "api"]

# ============================================================
# Mypy
# ============================================================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ============================================================
# Setuptools
# ============================================================
[tool.setuptools.packages.find]
where = ["src"]
